+++ addons/taliban/Veterans/CfgVehicles_Turrets_Veterans.hpp
@@ -1,90 +1,14 @@
-class TACU_Taliban_Veterans_V_I_AGS30: TACU_Assets_AGS_base {
+class TACU_Taliban_Veterans_V_I_ZU23: TACU_Assets_ZU23_Base {
     MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Taliban_Veterans_I";
+    faction = "TACU_Taliban_Militia_I";
     side = 2;
-    crew = "TACU_Taliban_Veterans_U_I_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_AGS30: TACU_Taliban_Veterans_V_I_AGS30 {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-};
-
-class TACU_Taliban_Veterans_V_I_D30: TACU_Assets_D30_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Taliban_Veterans_I";
-    side = 2;
-    crew = "TACU_Taliban_Veterans_U_I_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_D30: TACU_Taliban_Veterans_V_I_D30 {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-};
-
-class TACU_Taliban_Veterans_V_I_DSHKM: TACU_Assets_DSHKM_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Taliban_Veterans_I";
-    side = 2;
-    crew = "TACU_Taliban_Veterans_U_I_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_DSHKM: TACU_Taliban_Veterans_V_I_DSHKM {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-};
-
-class TACU_Taliban_Veterans_V_I_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Taliban_Veterans_I";
-    side = 2;
-    crew = "TACU_Taliban_Veterans_U_I_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_DSHKM_Mini: TACU_Taliban_Veterans_V_I_DSHKM_Mini {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-};
-
-class TACU_Taliban_Veterans_V_I_KORD: TACU_Assets_KORD_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Taliban_Veterans_I";
-    side = 2;
-    crew = "TACU_Taliban_Veterans_U_I_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_KORD: TACU_Taliban_Veterans_V_I_KORD {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-};
-
-class TACU_Taliban_Veterans_V_I_KORD_Mini: TACU_Assets_KORD_Mini_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Taliban_Veterans_I";
-    side = 2;
-    crew = "TACU_Taliban_Veterans_U_I_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_KORD_Mini: TACU_Taliban_Veterans_V_I_KORD_Mini {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-};
-class TACU_Taliban_Veterans_V_O_ZU23: TACU_Assets_ZU23_Base {
-    faction = "TACU_Taliban_Veterans_O";
-    side = 0;
-    crew = "TACU_Taliban_Veterans_U_O_Driver";
-    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
     textureList[] = {"Sand", 1};
+};
+class TACU_Taliban_Veterans_V_O_ZU23: TACU_Taliban_Veterans_V_I_ZU23 {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
 };
