+++ addons/taliban/config.cpp
@@ -20,6 +20,7 @@ class CfgPatches {
             "TACU_Taliban_Militia_U_I_Machinegunner",
             "TACU_Taliban_Militia_U_I_AT",
             "TACU_Taliban_Militia_U_I_Driver",
+            // Vehicles - Independent
             "TACU_Taliban_Militia_V_I_Kamaz_ZU23",
             "TACU_Taliban_Militia_V_I_Kamaz_Transport",
             "TACU_Taliban_Militia_V_I_Kamaz_Covered_Transport",
@@ -27,12 +28,9 @@ class CfgPatches {
             "TACU_Taliban_Militia_V_I_Offroad_Unarmed",
             "TACU_Taliban_Militia_V_I_Offroad_HMG",
             "TACU_Taliban_Militia_V_I_Offroad_AT",
-            "TACU_Taliban_Militia_V_I_AGS30",
-            "TACU_Taliban_Militia_V_I_D30",
-            "TACU_Taliban_Militia_V_I_DSHKM",
-            "TACU_Taliban_Militia_V_I_DSHKM_Mini",
-            "TACU_Taliban_Militia_V_I_KORD",
-            "TACU_Taliban_Militia_V_I_KORD_Mini",
+            // Turrets - Independent
+            "TACU_Taliban_Militia_V_I_ZU23",
+            // Units - Opfor
             "TACU_Taliban_Militia_U_O_Rifleman_01",
             "TACU_Taliban_Militia_U_O_Rifleman_02",
             "TACU_Taliban_Militia_U_O_Rifleman_03",
@@ -44,6 +42,7 @@ class CfgPatches {
             "TACU_Taliban_Militia_U_O_Machinegunner",
             "TACU_Taliban_Militia_U_O_AT",
             "TACU_Taliban_Militia_U_O_Driver",
+            // Vehicles - Opfor
             "TACU_Taliban_Militia_V_O_Kamaz_ZU23",
             "TACU_Taliban_Militia_V_O_Kamaz_Transport",
             "TACU_Taliban_Militia_V_O_Kamaz_Covered_Transport",
@@ -51,12 +50,7 @@ class CfgPatches {
             "TACU_Taliban_Militia_V_O_Offroad_Unarmed",
             "TACU_Taliban_Militia_V_O_Offroad_HMG",
             "TACU_Taliban_Militia_V_O_Offroad_AT",
-            "TACU_Taliban_Militia_V_O_AGS30",
-            "TACU_Taliban_Militia_V_O_D30",
-            "TACU_Taliban_Militia_V_O_DSHKM",
-            "TACU_Taliban_Militia_V_O_DSHKM_Mini",
-            "TACU_Taliban_Militia_V_O_KORD",
-            "TACU_Taliban_Militia_V_O_KORD_Mini",
+            // Turrets - Opfor
             "TACU_Taliban_Militia_V_O_ZU23",
             // Taliban Veterans
             "TACU_Taliban_Veterans_U_I_Rifleman_01",
@@ -70,6 +64,7 @@ class CfgPatches {
             "TACU_Taliban_Veterans_U_I_Machinegunner",
             "TACU_Taliban_Veterans_U_I_AT",
             "TACU_Taliban_Veterans_U_I_Driver",
+            // Vehicles - Independent
             "TACU_Taliban_Veterans_V_I_Kamaz_ZU23",
             "TACU_Taliban_Veterans_V_I_Kamaz_Transport",
             "TACU_Taliban_Veterans_V_I_Kamaz_Covered_Transport",
@@ -80,12 +75,9 @@ class CfgPatches {
             "TACU_Taliban_Veterans_V_I_Offroad_Unarmed_UP",
             "TACU_Taliban_Veterans_V_I_Offroad_HMG_UP",
             "TACU_Taliban_Veterans_V_I_Offroad_AT_UP",
-            "TACU_Taliban_Veterans_V_I_AGS30",
-            "TACU_Taliban_Veterans_V_I_D30",
-            "TACU_Taliban_Veterans_V_I_DSHKM",
-            "TACU_Taliban_Veterans_V_I_DSHKM_Mini",
-            "TACU_Taliban_Veterans_V_I_KORD",
-            "TACU_Taliban_Veterans_V_I_KORD_Mini",
+            // Turrets - Independent
+            "TACU_Taliban_Veterans_V_I_ZU23",
+            // Units - Opfor
             "TACU_Taliban_Veterans_U_O_Rifleman_01",
             "TACU_Taliban_Veterans_U_O_Rifleman_02",
             "TACU_Taliban_Veterans_U_O_Rifleman_03",
@@ -97,6 +89,7 @@ class CfgPatches {
             "TACU_Taliban_Veterans_U_O_Machinegunner",
             "TACU_Taliban_Veterans_U_O_AT",
             "TACU_Taliban_Veterans_U_O_Driver",
+            // Vehicles - Opfor
             "TACU_Taliban_Veterans_V_O_Kamaz_ZU23",
             "TACU_Taliban_Veterans_V_O_Kamaz_Transport",
             "TACU_Taliban_Veterans_V_O_Kamaz_Covered_Transport",
@@ -107,12 +100,7 @@ class CfgPatches {
             "TACU_Taliban_Veterans_V_O_Offroad_Unarmed_UP",
             "TACU_Taliban_Veterans_V_O_Offroad_HMG_UP",
             "TACU_Taliban_Veterans_V_O_Offroad_AT_UP",
-            "TACU_Taliban_Veterans_V_O_AGS30",
-            "TACU_Taliban_Veterans_V_O_D30",
-            "TACU_Taliban_Veterans_V_O_DSHKM",
-            "TACU_Taliban_Veterans_V_O_DSHKM_Mini",
-            "TACU_Taliban_Veterans_V_O_KORD",
-            "TACU_Taliban_Veterans_V_O_KORD_Mini",
+            // Turrets - Opfor
             "TACU_Taliban_Veterans_V_O_ZU23"
         };
         weapons[] = {
