+++ addons/taliban/subconfig_cup/script_component.hpp
@@ -1,0 +1,4 @@
+#define COMPONENT taliban_cup
+#define COMPONENT_BEAUTIFIED Taliban CUP
+#include "\x\tacu\addons\main\script_mod.hpp"
+#include "\x\tacu\addons\main\script_macros.hpp"
